Implement Scenario Play Tracking
A new UI element was added to i_menu.res, called NMPG-UCH. Its name is
based off the red checkbox element, NMPG-RCH, and it is simply an
unchecked checkbox.

NMPG-UCH is drawn when a scenario has not been played before and NMPG-RCH
is drawn when it has. Determining whether or not a scenario has been
played comes from scn_list.txt, placed in the user's config directory
(same place as save-game files).

If this file is missing, it is created dynamically, defaulting to unplayed
for all scenarios.

The checkboxes in the UI are non-functional; to reset the status, a user
can either delete scn_list.txt or manually edit it for the specific
scenario, 0=unplayed and 1=played.