Add global SaveGameInfo in AM.cpp
Rather than re-re-write all the game save/load code, we just keep a copy
of the last SaveGameInfo.

The primary purpose of this is to be able to keep a copy of the header's
internal file_name (now unioned with game_name), which will tell us if
the current game is a scenario, so we can track scenario completion.